Quite a bit of changes to structure, moving a lot to Core, and moving the source out of www (only the package go in there).
Changed a lot of the UI widgets to match the backbone pattern of having $el be a jQuery selector for the UI element.  Also added a good bit of inheritance to clean up common code in widgets.

Created the concept of a form and form widgets, with validation, etc.  Used this to create a form for personal info.  Created that form and adjusted the router.