Updated CollectionList and made an extension for selecting things by scrolling, with momentum, like a spinner.
Made a concept of a Modal Form that can be launched by a  widget.  Still in beginning phases with the number spinner.

A few race conditions still need to be cleaned up.  Occasionally the app freezes on load, and frequently it freezes while bringing up the ColorBox (modal dialog) with the number spinner.  Should see if promisifying more UI stuff fixes this.

Food CollectionList probably doesn't work right now the way it stores values.  Should be an easy fix.